사원 생성 구현 완료(이미지만 남음)

+++ src/main/java/com/kh/coreflow/chatting/model/websocket/StompHandler.java
@@ -57,7 +57,7 @@ public class StompHandler implements ChannelInterceptor {
         				    .map(SimpleGrantedAuthority::new)
         				    .collect(Collectors.toList());
         			// 4) 권한에서 부서코드 추출
-        			Long depId = jwtProvider.getDeptcode(jwtToken);
+        			int depId = jwtProvider.getDeptcode(jwtToken);
         			
         			UserDeptcode principal =  UserDeptcode.builder()
         			        .userNo(userNo)
